增加 Rasterfall Windows 单文件构建支持

+++ include/toy_renderer.h
@@ -41,7 +41,7 @@ struct toy_raster_cmd {
     int light;
     int fog;
     const struct toy_texture_view *texture;
-    long area;
+    long long area;
     int bbox_minx;
     int bbox_maxx;
     int bbox_miny;
